Added trk::Hit + fixed unsigned short ID

+++ Detectors/Upgrades/ALICE3/TRK/base/include/TRKBase/GeometryTGeo.h
@@ -101,7 +101,7 @@ class GeometryTGeo : public o2::detectors::DetMatrixCache
   /// \param int halfstave The half stave number for MLOT. Can be 0 or 1
   /// \param int module The module number for MLOT, from 0 to 10 (or 20)
   /// \param int chip The chip number for MLOT, from 0 to 8
-  int getChipIndex(int subDetID, int petalcase, int disk, int lay, int stave, int halfstave, int mod, int chip) const;
+  unsigned short getChipIndex(int subDetID, int petalcase, int disk, int lay, int stave, int halfstave, int mod, int chip) const;
 
   /// This routine computes the chip index number from the subDetID, volume, layer, stave, half stave, module, chip
   /// \param int subDetID The subdetector ID, 0 for VD, 1 for MLOT
@@ -111,7 +111,7 @@ class GeometryTGeo : public o2::detectors::DetMatrixCache
   /// \param int halfstave The half stave number for MLOT. Can be 0 or 1
   /// \param int module The module number for MLOT, from 0 to 10 (or 20)
   /// \param int chip The chip number for MLOT, from 0 to 8
-  int getChipIndex(int subDetID, int volume, int lay, int stave, int halfstave, int mod, int chip) const;
+  unsigned short getChipIndex(int subDetID, int volume, int lay, int stave, int halfstave, int mod, int chip) const;
 
   /// This routine computes subDetID, petal, disk, layer, stave, half stave, module, chip, given the chip index number
   /// \param int index The chip index number, starting from 0
@@ -125,8 +125,8 @@ class GeometryTGeo : public o2::detectors::DetMatrixCache
   /// \param int chip The chip number for MLOT, from 0 to 8
   bool getChipID(int index, int& subDetID, int& petalcase, int& disk, int& lay, int& stave, int& halfstave, int& mod, int& chip) const;
 
-  int getLastChipIndex(int lay) const { return mLastChipIndex[lay]; }
-  int getFirstChipIndex(int lay, int petalcase, int subDetID) const
+  unsigned short getLastChipIndex(int lay) const { return mLastChipIndex[lay]; }
+  unsigned short getFirstChipIndex(int lay, int petalcase, int subDetID) const
   {
     /// Get the first chip index of the active petal (VD) or layer (MLOT)
     if (subDetID == 0) { // VD
@@ -190,9 +190,9 @@ class GeometryTGeo : public o2::detectors::DetMatrixCache
   // std::vector<int> mNumberOfChipsPerStave;     ///< number of chips per stave in ML/OT
   // std::vector<int> mNumberOfChipsPerHalfStave; ///< number of chips per half stave in ML/OT
   // std::vector<int> mNumberOfChipsPerModule; ///< number of chips per module in ML/OT
-  std::vector<int> mLastChipIndex;     ///< max ID of the detctor in the petal(VD) or layer(MLOT)
-  std::vector<int> mLastChipIndexVD;   ///< max ID of the detctor in the layer for the VD
-  std::vector<int> mLastChipIndexMLOT; ///< max ID of the detctor in the layer for the MLOT
+  std::vector<unsigned short> mLastChipIndex;     ///< max ID of the detctor in the petal(VD) or layer(MLOT)
+  std::vector<unsigned short> mLastChipIndexVD;   ///< max ID of the detctor in the layer for the VD
+  std::vector<unsigned short> mLastChipIndexMLOT; ///< max ID of the detctor in the layer for the MLOT
 
   std::array<char, MAXLAYERS> mLayerToWrapper; ///< Layer to wrapper correspondence, not implemented yet
